restored component after conflict pull
