+++ src/pages/create-or-edit-question/CreateOrEditQuestion.styles.ts
@@ -9,6 +9,14 @@ const inputFontSize = {
 }
 
 export const styles = {
+  addButton: {
+    justifyContent: 'flex-start',
+    pl: '10px',
+    gap: '5px'
+  },
+  divider: {
+    my: '5px'
+  },
   input: {
     style: { padding: 0 }
   },
